Move gAudioErrorCallback call from af client to aps client
The gAudioErrorCallback will trigger B2G to restore the volume config to aps.
But the gAudioErrorCallback is send from af binder died in current code.

In a particular situation, like the 3rd mediaserver death in the following log,
af deatch notification is sent first, this will trigger the volume restore. But
the restore value is set in a dieing aps. This will result in no sound after
the new mediaserver up, since there will not be restore operation.

1st time, aps died first:
  09-09 12:29:14.529  1163  2375 W AudioSystem: AudioPolicyService server died!
  09-09 12:29:14.529  1163  2375 W AudioSystem: AudioFlinger server died!
2nd time, aps died first:
  09-09 12:31:24.449  1163  1168 W AudioSystem: AudioPolicyService server died!
  09-09 12:31:24.449  1163  1168 W AudioSystem: AudioFlinger server died!
3rd time, af died first:
  09-09 12:39:32.229  1163  2375 W AudioSystem: AudioFlinger server died!
  09-09 12:39:33.239  1163  1168 W AudioSystem: AudioPolicyService server died!

Change-Id: Ice5b6ddd2c86d3a61086aaaa604ce541e13f698d